Bubble Sort (Asc. Order) v.02
Improved bubble sort (efficient for larger arrays) | After the first pass, the largest number is guaranteed to be in the highest-numbered element of the array. After the second pass, the two highest numbers are “in place,” and so on. Instead of making nine comparisons on every pass, bubble sort is modified to make eight comparisons on the second pass, seven on the third pass and so on. Plus, the data in the array may already be in the proper order or near-proper order, so why make nine passes if fewer will suffice? The sort has been modified to check at the end of each pass if any swaps have been made. If none have been made, then the data must already be in the proper order, so the program should terminate. If swaps have been made, then at least one more pass is needed.